implement_label_count_totalization#53
